agent/pyproject.toml
[project]
name = "rfp-quest-agent"
version = "0.1.0"
description = "LangGraph agents for RFP.quest tender analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # CopilotKit
    "copilotkit>=0.1.65",

    # LangChain/LangGraph
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-google-genai>=2.0",
    "langchain-openai>=0.2",

    # FastAPI
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",

    # Web Search
    "tavily-python>=0.5",

    # Database
    "asyncpg>=0.30",
    "psycopg[binary]>=3.2",

    # PDF Processing
    "pypdf>=5.0",
    "pdfplumber>=0.11",

    # Utils
    "pydantic>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "tools", "prompts"]
include = ["main.py"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
